backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
pydantic==1.10.14
email-validator
python-multipart
bcrypt==4.0.1
